+++ cypress/e2e/EnrollmentEditEventPage/EnrollmentEditEventPageForm/EnrollmentEditEventPageForm.js
@@ -1,5 +1,17 @@
-import { defineStep as And, Given, Then, When } from '@badeball/cypress-cucumber-preprocessor';
+import { defineStep as And, Given, Then, When, Before } from '@badeball/cypress-cucumber-preprocessor';
 import { getCurrentYear } from '../../../support/date';
+
+Before({ tags: '@with-restore-event-schedule-date' }, () => {
+    cy.buildApiUrl('tracker', 'events/RIrfCcEP8Uu')
+        .then(url => cy.request(url))
+        .then((apiResponse) => {
+            const event = apiResponse.body;
+            const eventToUpdate = { ...event, scheduledAt: `${getCurrentYear() - 15}-01-07` };
+            return cy
+                .buildApiUrl('tracker?async=false&importStrategy=UPDATE')
+                .then(eventUrl => cy.request('POST', eventUrl, { events: [eventToUpdate] }));
+        });
+});
 
 const changeEnrollmentAndEventsStatus = () => (
     cy.buildApiUrl(
@@ -84,7 +96,7 @@ When(/^the user clicks on the cancel button/, () =>
 When(/^the user set the apgar score to (.*)/, score =>
     cy
         .get('[data-test="widget-enrollment-event"]')
-        .find('[data-test="capture-ui-input"]')
+        .find('input[type="text"]')
         .eq(2)
         .clear()
         .type(score)
@@ -118,8 +130,8 @@ When('the user clicks switch tab to Schedule', () => {
 
 Then('the user selects another schedule date', () => {
     cy.get('[data-test="schedule-section"]').within(() => {
-        cy.get("[data-test='capture-ui-input']").eq(0).should('have.value', `${getCurrentYear() - 15}-01-07`);
-        cy.get("[data-test='capture-ui-input']").eq(0)
+        cy.get('input[type="text"]').eq(0).should('have.value', `${getCurrentYear() - 15}-01-07`);
+        cy.get('input[type="text"]').eq(0)
             .clear()
             .type(`${getCurrentYear()}-08-01`)
             .blur();