+++ cypress/e2e/WidgetsForEnrollmentPages/WidgetsForEnrollmentAddEventPage/WidgetsForEnrollmentAddEventPage.js
@@ -6,8 +6,8 @@ import '../WidgetTab';
 
 Then('you can assign a user when scheduling the event', () => {
     cy.get('[data-test="assignee-section"]').within(() => {
-        cy.get('[data-test="capture-ui-input"]').click();
-        cy.get('[data-test="capture-ui-input"]').type('Tracker demo');
+        cy.get('input[type="text"]').click();
+        cy.get('input[type="text"]').type('Tracker demo');
         cy.contains('Tracker demo User').click();
     });
     cy.get('[data-test="assignee-section"]').within(() => {