+++ src/core_modules/capture-core-utils/validators/form/index.js
@@ -1,7 +1,5 @@
 // @flow
 export { hasValue } from './compulsory.validator';
-export { isValidDate } from './date.validator';
-export { isValidDateTime } from './dateTime.validator';
 export { isValidEmail } from './email.validator';
 export { isValidInteger } from './integer.validator';
 export { isValidPositiveInteger } from './integerPositive.validator';
@@ -11,7 +9,6 @@ export { isValidNumber } from './number.validator';
 export { isValidPercentage } from './percentage.validator';
 export { isValidTime } from './time.validator';
 export { isValidUrl } from './url.validator';
-export { isValidAge } from './age.validator';
 export { isValidPhoneNumber } from './phone.validator';
 export { isValidOrgUnit } from './orgUnit.validator';
 export { isValidCoordinate } from './coordinate.validator';