+++ src/core_modules/capture-ui/AgeField/AgeField.component.js
@@ -26,11 +26,17 @@ type InputMessageClasses = {
 }
 
 type DateParser = (value: string) => { isValid: boolean, momentDate: any };
+
 type DateStringFromMomentFormatter = (momentValue: Object) => string;
+
+type ValidationOptions = {
+    error?: ?string,
+    errorCode?: ?string,
+};
 
 type Props = {
     value: ?AgeValues,
-    onBlur: (value: ?AgeValues) => void,
+    onBlur: (value: ?AgeValues, options: ?ValidationOptions) => void,
     onChange: (value: ?AgeValues) => void,
     onRemoveFocus: () => void,
     orientation: $Values<typeof orientations>,
@@ -141,7 +147,7 @@ class D2AgeFieldPlain extends Component<Props> {
         this.props.onBlur(calculatedValues);
     }
 
-    handleDateBlur = (date: ?string) => {
+    handleDateBlur = (date: ?string, options: ?ValidationOptions) => {
         const { onParseDate, onGetFormattedDateStringFromMoment, onRemoveFocus, moment } = this.props;
         onRemoveFocus && onRemoveFocus();
         const calculatedValues = date ? getCalculatedValues(
@@ -149,7 +155,7 @@ class D2AgeFieldPlain extends Component<Props> {
             onParseDate,
             onGetFormattedDateStringFromMoment,
             moment) : null;
-        this.props.onBlur(calculatedValues);
+        this.props.onBlur(calculatedValues, options);
     }
 
     renderMessage = (key: string) => {
@@ -200,17 +206,13 @@ class D2AgeFieldPlain extends Component<Props> {
             value,
             onBlur,
             shrinkDisabled,
-            dateCalendarOnConvertValueIn,
-            dateCalendarOnConvertValueOut,
             dateCalendarWidth,
-            datePopupAnchorPosition,
-            dateCalendarTheme,
-            dateCalendarLocale,
             datePlaceholder,
             moment,
             onParseDate,
             ...passOnProps
         } = this.props;
+
         const dateInputContainerClass = classNames(
             { [defaultClasses.ageDateInputContainerHorizontal]: !isVertical },
         );
@@ -222,11 +224,6 @@ class D2AgeFieldPlain extends Component<Props> {
                     value={currentValues.date}
                     onChange={date => onChange({ ...currentValues, date })}
                     calendarWidth={dateCalendarWidth}
-                    popupAnchorPosition={datePopupAnchorPosition}
-                    calendarTheme={dateCalendarTheme}
-                    calendarLocale={dateCalendarLocale}
-                    calendarOnConvertValueIn={dateCalendarOnConvertValueIn}
-                    calendarOnConvertValueOut={dateCalendarOnConvertValueOut}
                     placeholder={datePlaceholder}
                     {...passOnProps}
                 />